requirements.txt
django
django-rest-framework
django-rest-knox
